+++ smsapi/sms.go
@@ -45,7 +45,7 @@ type SmsResultCollection struct {
 
 type SmsResponse struct {
 	Id              string     `json:"id,omitempty"`
-	Points          float32    `json:"points,omitempty"`
+	Points          Points     `json:"points,omitempty"`
 	Number          string     `json:"number,omitempty"`
 	DateSent        *Timestamp `json:"date_sent,omitempty"`
 	SubmittedNumber string     `json:"submitted_number,omitempty"`